refactor(component): :recycle: use new methods in `inversify-hooks-esm`
move `container` from global into react context instead
