+++ .vscode/settings.json
@@ -16,7 +16,8 @@
         "titleBar.activeBackground": "#61dafb",
         "titleBar.activeForeground": "#15202b",
         "titleBar.inactiveBackground": "#61dafb99",
-        "titleBar.inactiveForeground": "#15202b99"
+        "titleBar.inactiveForeground": "#15202b99",
+        "commandCenter.border": "#15202b99"
     },
     "peacock.color": "#61dafb"
 }
