+++ next.config.js
@@ -13,6 +13,9 @@ const isProduction = process.env.NODE_ENV === 'production'
 const isNextExportCommand = process.env.NEXT_EXPORT_COMMAND === 'true'
 
 const nextConfigurations = {
+  // experimental: {
+  //   fallbackNodePolyfills: false,
+  // },
   images: {
     domains: ['www.floatui.com', 'images.clerk.dev'],
     ...(isNextExportCommand ? {
