+++ package.json
@@ -23,76 +23,76 @@
   },
   "dependencies": {
     "@abraham/reflection": "^0.10.0",
-    "@clerk/nextjs": "^3.8.1",
-    "@fortawesome/fontawesome-svg-core": "^6.1.2",
-    "@fortawesome/free-brands-svg-icons": "^6.1.2",
-    "@fortawesome/free-regular-svg-icons": "^6.1.2",
-    "@fortawesome/free-solid-svg-icons": "^6.1.2",
-    "@fullpage/react-fullpage": "^0.1.26",
-    "@linaria/babel-preset": "^4.1.1",
-    "@linaria/core": "^4.1.0",
-    "@linaria/react": "^4.1.1",
-    "@linaria/shaker": "^4.1.1",
-    "@linaria/webpack-loader": "^4.1.1",
-    "@popperjs/core": "^2.11.5",
-    "@prisma/client": "^4.1.1",
-    "@reduxjs/toolkit": "^1.8.3",
+    "@clerk/nextjs": "^4.3.2",
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-brands-svg-icons": "^6.2.0",
+    "@fortawesome/free-regular-svg-icons": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@fullpage/react-fullpage": "^0.1.28",
+    "@linaria/babel-preset": "^4.2.0",
+    "@linaria/core": "^4.1.2",
+    "@linaria/react": "^4.1.3",
+    "@linaria/shaker": "^4.2.0",
+    "@linaria/webpack-loader": "^4.1.3",
+    "@popperjs/core": "^2.11.6",
+    "@prisma/client": "^4.3.1",
+    "@reduxjs/toolkit": "^1.8.5",
     "axios": "^0.27.2",
     "clsx": "^1.2.1",
     "comlink": "^4.3.1",
     "compression-webpack-plugin": "^10.0.0",
     "cross-env": "^7.0.3",
-    "flowbite": "^1.5.1",
+    "flowbite": "^1.5.3",
     "flowbite-datepicker": "^1.2.2",
-    "flowbite-react": "^0.1.4",
+    "flowbite-react": "^0.1.11",
     "formik": "^2.2.9",
-    "inversify-esm": "^2.0.0",
-    "inversify-hooks-esm": "^2.0.0",
-    "inversify-props-esm": "^2.0.0",
-    "jose": "^4.8.3",
-    "next": "12.2.3",
+    "inversify-esm": "^2.0.1",
+    "inversify-hooks-esm": "^2.1.0",
+    "inversify-props-esm": "^2.0.1",
+    "jose": "^4.9.3",
+    "next": "12.3.1",
     "next-compose-plugins": "^2.2.1",
     "next-linaria": "^1.0.1-beta",
     "next-optimized-images": "^2.6.2",
-    "next-pwa": "^5.5.5",
-    "next-redux-cookie-wrapper": "^2.1.2",
-    "next-redux-wrapper": "^7.0.5",
+    "next-pwa": "^5.6.0",
+    "next-redux-cookie-wrapper": "^2.2.1",
+    "next-redux-wrapper": "^8.0.0",
     "purge-fontawesome": "^1.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.2",
     "react-popper": "^2.3.0",
-    "react-redux": "^8.0.2",
-    "react-toastify": "^9.0.7",
-    "read-excel-file": "^5.4.4",
-    "sass": "^1.54.0",
-    "tailwindcss": "^3.1.7",
+    "react-redux": "^8.0.4",
+    "react-toastify": "^9.0.8",
+    "read-excel-file": "^5.5.0",
+    "sass": "^1.55.0",
+    "tailwindcss": "^3.1.8",
     "ts-node": "^10.9.1",
     "windicss": "^3.5.6",
-    "windicss-webpack-plugin": "^1.7.5",
+    "windicss-webpack-plugin": "^1.7.6",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.0.3",
-    "@commitlint/config-conventional": "^17.0.3",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-conventional": "^17.1.0",
     "@types/lodash-es": "^4.17.6",
-    "@types/node": "^18.6.3",
-    "@types/react": "^18.0.15",
-    "@typescript-eslint/eslint-plugin": "^5.31.0",
+    "@types/node": "^18.7.21",
+    "@types/react": "^18.0.21",
+    "@typescript-eslint/eslint-plugin": "^5.38.0",
     "add": "^2.0.6",
-    "autoprefixer": "^10.4.8",
-    "eslint": "8.20.0",
-    "eslint-config-next": "12.2.3",
+    "autoprefixer": "^10.4.12",
+    "eslint": "8.24.0",
+    "eslint-config-next": "12.3.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.1",
     "lint-staged": "^13.0.3",
     "lodash-es": "^4.17.21",
-    "postcss": "^8.4.14",
-    "postcss-import": "^14.1.0",
+    "postcss": "^8.4.16",
+    "postcss-import": "^15.0.0",
     "prettier": "^2.7.1",
-    "prisma": "^4.1.1",
-    "typescript": "^4.7.4",
+    "prisma": "^4.3.1",
+    "typescript": "^4.8.3",
     "yarn": "^1.22.19"
   },
   "packageManager": "yarn@3.2.2",
