+++ src/components/03-organisms/header/mosque/Base.tsx
@@ -1,10 +1,11 @@
 import { type MouseEvent } from 'react'
-import { container } from 'inversify-hooks-esm'
+import { useContainerGet } from 'inversify-hooks-esm'
 import { useUser } from '@clerk/nextjs'
 import { createPopper } from '@popperjs/core'
 import { useSelector } from 'react-redux'
 
-import Image from 'next/image'
+// import Image from 'next/image'
+import Image from 'next/future/image'
 import Link from 'next/link'
 
 import { CONTROLLER_USER_ADMIN_LOGOUT_BASE_CLIENT, IControllerCoreLogout } from 'controllers'
@@ -35,7 +36,7 @@ const HeaderMosqueBase = <
   const qurbanEventYearExecution = useSelector(getQurbanEventYearExecution)
   const qurbanEventId = useSelector(getQurbanEventId)
 
-  const logoutCtrl = container.get<IControllerCoreLogout>(
+  const logoutCtrl = useContainerGet<IControllerCoreLogout>(
     CONTROLLER_USER_ADMIN_LOGOUT_BASE_CLIENT
   )
 
@@ -128,8 +129,10 @@ const HeaderMosqueBase = <
                       alt={`${user.fullName} Profile Image URL`}
                       className="w-full rounded-full align-middle border-none shadow-lg"
                       src={user.profileImageUrl}
-                      layout='fill'
+                      // layout='fill'
                       priority
+                      width={48}
+                      height={48}
                     />
                   </span>
                 </div>
