+++ src/controllers/mosque/admin/find/get-resource-data.client.hooks.ts
@@ -1,4 +1,4 @@
-import { container } from 'inversify-hooks-esm'
+import { useContainerGet } from 'inversify-hooks-esm'
 import { useState } from 'react'
 import { AxiosInstance } from 'axios'
 
@@ -19,7 +19,7 @@ export function useControllerMosqueAdminFindGetResourceDataClient<
   const [loading, setLoading] = useState(false)
 
   const { transformRequestBody } = useControllerMosqueFindTransformRequestBodyClient()
-  const remoteService = container.get<AxiosInstance>(SERVICE_CORE_REMOTE_BASE)
+  const remoteService = useContainerGet<AxiosInstance>(SERVICE_CORE_REMOTE_BASE)
 
   const getData = async () => {
     setLoading(true)
