+++ src/controllers/qurban-event/admin/find/get-resource-data.client.hooks.ts
@@ -1,4 +1,4 @@
-import { container } from 'inversify-hooks-esm'
+import { useContainerGet } from 'inversify-hooks-esm'
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { AxiosInstance } from 'axios';
@@ -31,7 +31,7 @@ export function useControllerQurbanEventAdminFindGetResourceDataClient<
   const fallbackQurbanEventId = useSelector(getQurbanEventId)
   const qurbanEventDataIsEmpty = useSelector(checkQurbanEventDataEmpty)
 
-  const remoteService = container.get<AxiosInstance>(SERVICE_CORE_REMOTE_BASE)
+  const remoteService = useContainerGet<AxiosInstance>(SERVICE_CORE_REMOTE_BASE)
 
   const getData = async (options: OptionType) => {
     const mosqueId = options?.mosqueId ?? fallbackMosqueId
